PRT-839 When refreshing tree view, the whole tree must be refresh
When actions are performed, refreshListing from useGalleryListing is invoked.
This causes the main listing to be refreshed and the UI updated to reflect the
new state of the files. This works just fine in grid and carousel view where
there is just the one listing, but in tree view there is a hierarchy of
listings and if the user performs an edit on a file deeply nested in some
folders we were not updating that part of the UI as only the root listing was
being refreshed.

This changes does two things to rectify this.
  1. refreshListing sets and unsets the loading flag, resulting in the tree
     components being unmounted. This would be sufficient, but all of the tree
     nodes that were expanded would now be collapsed. So,
  2. we move the unbundling of the FetchingData object to inside the TreeView
     component, thereby ensuring that whilst the tree is unmounted the state
     stored in the expandedItems state variables are not lost.

As such, when the root listing is refreshed, the tree is unmounted and
re-mounted, the nodes that were open are expanded and their contents
re-fetched, and the whole process propagates down the tree.

This is not the most efficient way to update the UI, especially if the user is
working with very deeply nested folder hierarchies but it does keep the code
relatively simple. The actions menu does not need to know anything about how
the current listing is being displayed to the user; just that a number of files
have been selected and that the whole listing should be refreshed once the
action completes successfully.

We also maintain the ability for actions to be performed on files in different
folders at different levels at the same time as the logic is not dependent on
keeping track precisely which folders need to be refreshed -- we're just
refreshing everything. If a folder contains multiple pages of files then all of
the pages will be refreshed and the propagating logic continues from there too.