Differentiate loading and refreshing in new gallery listing
This changes adds a new refreshing state, alongside the standard and empty
listings, so that when the listing is being refreshed we don't have to use the
loading state to ensure that the UI is updated correctly. This new state allows
us to continue to show the old listing state whilst the new one is being
fetched, resulting in a more seamless UI without white flashes being caused by
mostly identical listings disappearing and reappearing.