Various tweaks to refreshing of listings in new gallery (#221)
After the user has performed an action, we need to refresh the UI to
show the new state of their files and folders. Commit 8d05363d4d ensured
that the refreshListing function set and unset the loading flag, thereby
causing the listing components to be unmounted and remounted; which
would result in all of the tree view nodes fetching their state again.

This, however, had a not-insignificant downside: in grid view, the
listing would vanish and then re-appear after the refresh. Given that
after any particular action most of the listing will remain unchanged
(with just the odd file being deleted, duplicated, renamed, etc), it
would be better if the content stayed visible and only updated once the
new listing had been fetched.

This change fixes this by introducing a new flag to differentiate when a
listing is being fetched for the first time (still using `loading`)
versus being refreshed. As such, the UI components can keep showing the
previous listing until the new one has been fetched. Tree view watches
for the changes to this `refreshing` flag and triggers `refreshListing`
on each of its sub-listings when the flag changes from false to true.

This in turn caused another issue: when deleting a folder, tree view
would attempt to refresh its contents before the parent folder would
remove it from the DOM. We make no attempt to prevent this network
request, but instead catch it -- expecting that any attempt to refresh a
listing that fails with an error message that mentions a **deleted
folder** is just fine and should be ignored.

There are further minor issues around selection not being persisted
across refreshes in tree view that are harder to solve whilst keeping
logic that is needed to clear the selection in other parts of the UI.
These cases are documented instead.