PRT-852 Prevent tree folder loading impacting selection
Whenever a listing is loaded it clears the selection, which makes sense as we
don't want the user to be able to select files that are no longer viewable.
However, this meant that whenever the user opened a folder of tree view we
would fetch its contents and clear the selection, preventing the user from
selecting a closed folder; they would have to open it first and then tap it
again to select it.

By wrapping all of the contents of the Tree in a GallerySelection, we prevent
the selection clearing on the fetch of each folder from clearing the selection
of the whole tree.