Fix minor offset from buttons that open menus
When a button that opens a menu is tapped, the menu should open right next to
the boundary of the button to reinforce the semantic relationship between the
button and its label and the contents of the button. Whilst this was generally
happening correctly, the offset was being calculated from the particular DOM
node that was tapped, rather than the button in question. If the user happened
to tap the icon or text label inside the button then the menu would open with
respect to it and the menu would be positioned slightly off. The correct thing
to do is to use the `currentTarget` rather than the `target`, which is the
actual DOM node that the event listener is attached to rather than the node
from which the event originates.