Fix isBeingMoved logic and extract to function
In the process of extract isBeingMoved into a separate function to reduce the
cyclomatic complexity of MainPanel, I noticed that the logic isn't quite right
because when the fileBeingMoved is the file in question we were comparing a
reference to the file to the id of the file. As such, if you dragged a folder
without first selecting it or anything else, it would continue to operate as a
dropzone for itself.