RSDEV-386 Move new Gallery search out of app bar (#233)
With the app bar being rolled out across more of the product to provide
a unified navigation mechanism, including the search box in the app bar
on the new gallery page suggests that the search is applied to the whole
product. In reality, it is only applied to the current folder of the
current gallery section, so it is best placed outside of the app bar.