RSDEV-422 App bar navigation for new Gallery, Inventory, and Apps (#246)
This change adds an app bar to Inventory and Apps, and adds behaviour to
the new Gallery's app bar to facilitate mobile-friendly navigation
across the product. It brings the whole of the ELN into reach directly
from Inventory, it unifies the location of the help menu as per the
accessibility standards, and sets the grounding for how we want to bring
the entire product together into a more cohesive experience.

---------

Co-authored-by: matthias <matthias@researchspace.com>