Handle case where /gallery/folderId is the id of the section root
In the case where the folder id encoded in the URL is the id of the gallery
section root, we should set the path to the empty array, otherwise the gallery
section root shows up twice in the breadcrumbs.

This is shouldn't typically happen under normal usage, and it likely suggests
that the logic for generating links is faulty, but in the case where the
section root is encoded in the URL, we should handle it gracefully.