RSDEV-506 Filestores should not mutate URL
As the user opens folders, we mutate the URL so that they may use the browser's
back and forward buttons to navigate around. However, we should not be doing
this when traversing filestores as it is not possible to determine the full
path of the current folder, including which filestore it is in, solely from
information that can be easily encoded into the URL. As such, it does mean the
the back/forward buttons, bookmarking, etc will not work as expected when
navigating filestores.

It is not possible to add links to documents that refer to whole folders into
external filestores so there aren't any links in documents that will be broken
by this change.