Don't allow hiding new gallery banner
Whilst being able to hide the banner was a nice idea, it kept coming back as
the user navigates around the gallery as many of the operations, such as
opening folders and tapping the sidebar, count as navigation events.

Moving the app bar and banner outside of the routing logic is not trivial
because of the sidebar toggle hamburger menu which requires that its state be
held by a parent component of both of the app bar and sidebar and so would
require prop drilling through the small components that implement each branch
of the routing logic. We probably want to do this anyway so that the whole app
bar, including branding logo, isn't loaded all the time, but this is a quicker
fix for the issue at hand.

This change then makes the banner smaller so that it doesn't take up quite so
much screen real estate, especially on mobile devices, by removing a
superfluous sentence.

Finally, a link is made to the gallery documentation.