PRT-877 Remember url of branding banner image in session storage
This speeds up the time to render the branding banner image when the user
navigates to a different page, regardless of whether the actual image data is
cached by the browser or not.

A better long-term solution will be to move the app-bar outside of the routing
logic so that as the user navigates around we don't need to re-render the app
bar at all, which will also resolve the janking rendering of the conditional
parts of the app bar. However, this is far more complicated to do whilst much
of the ELN is not based on react.