PRT-886 Fix label on Inventory container drag handle
When you click and drag over a grid or visual container in Inventory, a
rectangular region apears within which all locations are selected. Within this
region a label is meant to appear showing the number of locations selected.
This label was instead showing the total number of locations in container,
regardless of the number that would be selected.

This was caused by commit 012f6e72, which changed the way that functions like
`isShallow` access the current search. In the case of Dragger.js, `isShallow`
was not being called correctly, but rather the function object of the method
was itself being returned by the callback passed to `filter` when calculating
the number of locations that would be selected. The function object was
silently coerced to boolean as it is truthy and the filter became a no-op.

The only thing I can think of that would have caught this would have been a
custom utility function reimplementing `Array.prototype.filter` that would have
thrown an error if the callback returned a non-boolean value; but even then
you'd have to remember to use the utility function instead of the built-in
method everytime.