Clean up TitledBox component
This change primarily removes the dependency on UiStore. The isSmall
conditional was used to remove the padding on smaller viewports but the UI ends
up looking rather cramped. By removing the dependency we make this component
easier to migrate to TypeScript and as I intend to use it on new Inventory
pages in the not-too-distant future, I can avoid further spreading the global
state.