RSDEV-549 Disable spellcheck on username fields
In Safari on macOS, textfields automatically apply spelling correcting to
entered text, which can be particularly annoying in username fields as
usernames do not resemble regular text.

Spellcheck is already disabled for most username fields across the system as
they are typically select widgets, allowing the user to filter a list of
available options. However, there are a couple of places where the user is
expect to enter a username manually or where the filtering mechanism is
independent of the textfield so the default spellcheck behavior does not get
applied. In both those cases, this change disables the spellcheck.