Organise code that operates on units and their utility functions. (#285)
Everything relating to units and measure how lives in
`src/stores/definitions/Units.js`, which means that `src/util/*.js`
files do not have to depend on code outside of `src/util/`. This fixes a
violation of a dependency cruiser rule, and makes it easier to migrate
the code to TypeScript in a piecemeal fashion as the modules aren't so
tightly coupled together. This change also adds documentation and
removes some unused code.