Add TypeScript to the project (#271)
This change adds TypeScript to the codebase, making it work
alongside Flow, Jest, Eslint, and the various other tools that we use,
with the ultimate aim being to move the codebase from Flow to TypeScript
in small pieces when we can find the time.

This migration works because both webpack and jest can transpile the
flow and typescript code into javascript and then have the former call
the later where necessary. To make the type check continue to work, the
boundary between the flow js code at the root of the module tree and
typescript at the leaves is annotated with a flow library declaration
file so that the calling flow code can get the right types.

Also includes a few files; test subjects that were necessary in
identifying all of the changes necessary to all of the config files to
get everything working: some simple components, a test file, and a
cyclical dependency.