RSDEV-586 Organise static assets relating to logos and brand colours (#290)
This makes is easier to migrate the codebase to TypeScript, as there are
now fewer dependencies across the different pages of the product.
Previously, several parts of the code base depended on the code for the
apps page as that was where the logos and brand colours were first used,
but this meant that migrating any of those other parts could only be
done by first migrating the apps page. Whilst I could have just migrated
the apps page before the others, I thought it better to take this
opportunity to tidy up the codebase, remove these cross dependencies,
and reduce the number of dependency cruiser warnings.