RSDEV-570 Add aria-current support to AccentMenuItem (#298)
This change adds support for aria-current to the AccentMenuItem
component, utilising the prop in the nav app bar and the gallery to
enhance the accessibility of the menus that use the AccentMenuItem
component.

This change was pulled out of RSDEV-570, as it requires the use of the
current marker for menus.