RSDEV-518 Add support for IGSNs to print action (#294)
In Inventory, we have a mechanism of printing a record's QR code onto a
label. This change re-uses this functionality for printing QR codes to
IGSN DOIs in layout analogous to raffle books, wherein each IGSN is
printed twice to aid with offline record keeping of used IGSNs.

This change
1. On all print outs, the Global Id of the parent container (labelled
"Location") is replaced with the URL of the container, as that would be
more useful on a printout.
2. Splits the functionality that allows for the printing of all of the
barcodes associated with a particular record (figure 1) from that of the
functionality for printing the first barcode associated with each record
(figure 2). This is because the latter can be re-used for IGSNs given
the 1-to-1 mapping between Inventory records and barcodes.
3. Additional options when printing for choosing Global ID or IGSN, and
1 or 2 copies of each label. Together, these allow for IGSNs to be
printed in a raffle-book like format.

Figure 1:
<img width="278" alt="image"
src="https://github.com/user-attachments/assets/dbeb9e28-940d-4bc9-8136-e5e5bc0202c1"
/>

Figure 2:
<img width="396" alt="image"
src="https://github.com/user-attachments/assets/f9b36b27-9a14-4e10-839a-73c183711703"
/>