PRT-907 Synchronise SearchStore listing when creating IGSN
When we edit a record on the right panel, we update the listing in the left
panel so that they continue to point to the same object in memory. This is so
that the contents of the listing table and logic performed by the context
actions is consistent. However, we were not doing this when creating a new
IGSN, resulting in the new logic for printing IGSNs thinking that the record
did not have an IGSN.