RSDEV-570 A new Inventory page for managing IGSNs (#300)
Including bulk registering, deleting, and printing. This change also
includes a TinyMCE plugin for quickly adding a table of IGSNs to a
document.

We're merging the code in now to avoid a long-lived branch, but the UI
is hidden from view until we implement the changes to make the
unassigned IGSNs useful