Webpack should support JSON imports without extension
Jest already is configured to import JSON files without requiring that the
extension be specified (see package.json), but Webpack was not configured to do
so. This is fine as it seems that we only import JSON files in the tests, but
dependency cruiser uses webpack to walk the dependency graph and as such it was
not able to resolve the JSON imports in the test source code. This commit adds
the configuration to allow JSON imports without the extension, thereby allowing
dependency cruiser to verify the use of JSON imports.

I found the issue because my script for identifying JavaScript files that can
be migrated to TypeScript was failing to identify that some jest tests could be
migrated because the extension-less JSON imports were not being resolved.