Rsdev 581 merge chemistry changes to open source (#323)
## Description ##
This PR copies the open-source chemistry replacement changes from the
closed-source repo.

Chemistry will still be disabled by default as the users will need to
have an instance of the chemistry service running and populate the
required properties which is described in
`DevDocs/public/RSpaceConfiguration.md`. I'll update
`RspaceConfiguration.md` separately and add the 2 other chemistry
properties for indexing search chems and migrating images, and with a
link to the chemistry service repo.

## Testing notes
- jenkins build with all front end and back end tests passes
https://jenkins.researchspace.com/job/rspace-web/job/rsdev-581-merge-chemistry-changes-to-open-source/2/
- Most of the integration tests which require a chemistry service were
marked to run as "nightly", although some weren't, or were marked as
nightly but not run inside a `ConditionalTestRunner` so would fail when
there's no access to the howler instance. I marked a couple test as
nightly or added the `ConditionalTestRunner` so that the tests pass
whether running inside the rspace vpn or not.
- I'll create a tech debt ticket to remove all of the references to
"howler" in the code, such as in `@PropertySource` annotations. This
doesn't mean anything outside of rspace infra and should be managed via
property files. It would also be good to clear up what is required for
running the "nightly" tests as it's quite unclear and not something
anyone in the open-source community could currently do.

---------

Co-authored-by: rs-fraser <fraser.macinnes>