PRT-919 Don't prohibit the user from entering a hyphen when the field is empty
The issue preventing negative temperature values is that the input handlers in
SpecifiedStorageTemperature.js immediately parse input with parseInt() as soon
as a character is typed. When a user types a hyphen as the first character,
parseInt("-", 10) returns NaN rather than allowing the user to complete typing
a negative number. For some reason we were applying parseInt twice, so this
change removes the first one so that the string value of the field can be
persisted allowing the user to type in hyphen and then a number.