Remove analytics event that never fires
`updateValue` will always be undefined because the return type of the `update`
method is void. As such, the if statement will never be true and analytics
event never recorded.

I also haven't the slightest clue why this event was added to SubSampleModel in
the first place, implying it was only meant to record succesful updates to
subsamples, but then the name nor the metadata of the event suggests this...?