RSDEV-633 Save chemicals in copied documents to the chemistry service (#332)
## Description ##
When a document is copied, the chemical element was saved via the dao
layer rather than manager layer, which meant the call to save the
chemical to rspace document pair to the chemistry service was bypassed.

This meant the copied document wouldn't be returned as a search hit when
searching for the chemical element.

The fix is to save the copied document via the manager rather than the
dao layer.

## Design decisions
As the test requires an instance of the chemistry service, it's been
ignored in the open-source build, which is the case for other
chemistry-dependent tests.

I've created https://researchspace.atlassian.net/browse/RSDEV-673 with
some ideas on how to handle these tests in a nicer way.

I'll enable this test in the closed-source repo once it's been synced.

## Testing notes
Passing Jenkins build here:
https://jenkins.researchspace.com/job/rspace-web/job/rsdev-633-save-copied-chems/2/
(shows there are no regressions, though the new test will be ignored in
this build.)

---------

Co-authored-by: rs-fraser <fraser.macinnes>
Co-authored-by: Maciej Kowalski <mKowalski256@users.noreply.github.com>