PRT-922 Finish removing removeScopedToasts
This change was intended to be completed by commit c9409a9 (RSPAC-2806) but was
not fully applied. This commit removes the remaining reference to
removeScopedToasts in the codebase as its inclusion prevented a visual
containers's location image from being modified twice in an editing session.

It's not at all clear why Flow did not find this reference 18 months age, but
has now due the changes taking place to migrate to TypeScript.