WebSocket POC for notifications.
This commit introduces WebSocket functionality to handle real-time notifications without the need for polling. It uses Spring websocket and adds a `NotificationController` for publishing messages, and a `WebSocketConfig` that configures a message broker.

The front-end changes mean that the notification count on the toolbar is no longer polled but instead the front-end code subscribes to a topic keyed on the users id. When a notification is generated for the user, the message count is published to the topic `topics/notifications/{userId}`.