Refactor movable mixin and RecordWithQuantity classes to be simpler and more type safe (#340)
## Description ##
Now that I've migrated Inventory's definitions, stores, and models to
TypeScript, I can actually utilise some of the TypeScript's capabilities
to enhance the type safety. This change replaces the `Movable` mixin
with a `HasLocation` one and replaces the `RecordWithQuantity` class
with a `HasQuantity` mixin. This simplifies the class hierarchy and
makes the extension of Inventory record classes more uniform

This change also removes the `classMixin` util function, replacing it
with the use of anonymous classes. The improvement being that TypeScript
can reason about the properties declared on the mixin class, it has its
own properties, and a constructor that can be used to initialise those
properties. We can begin using TypeScript's support for `private`,
`protected`, and `public` scoping.

## Design decisions
A new pattern I've introduced with this, is for the interface to require
a symbolic property, just like the language built-in Iterable protocol.
This means that runtime we can filter a collection of objects based on
their adherence to the interface.

## Testing notes
Regressions are possible, though most of the changes are at the
type-level.