PRT-934 Fix rich text field onChange handler
Commit c8ca95ae removed the `onChange={null}`, as I figured that it didn't do
anything. Turns out it did, because `props` includes `onChange` and `Editor`
accepts an `onChange` prop, even though we don't use it. We were then setting
`onChange` to `null`, so that the change handler being passed in would not be
called directly. I could have reverted that, but then it would be no clearer to
the next person why we were setting it to null. Instead, I'm pulling `onChange`
out of `props`, so its not passed to `Editor` at all.