Rename Inventory Result class (#346)
## Description ##
The abstract base class of all Inventory records has been called
"Result" since the beginning of the product, but this is confusing
because it has the same name as the newer Result util class (which is
itself named after the same functionality in languages like Rust and
Roc). Furthermore, "Result" is not intuitive to someone new coming to
the codebase as the class is used for more than just search results in
Inventory.