PRT-935 IGSN place description field loses focus (#358)
## Description ##
Whenever the text of the IGSN place text description field is changed,
it loses focus thanks to a react re-rendering. This change fixes this by
only causes the re-rendering when the user moves focus away from the
textfield. This is a little hacky but the whole IGSN UI code is a mess
and I don't fancy refactoring it all