PRT-936 Show alerts in Identifiers TinyMCE window (#363)
## Description ##
When the user performs an action, like deleting an IGSN, in the
Identifiers TinyMCE dialog then there should be success and error
alerts. Because the document editor is not one big react app, we have to
ensure that each island of react provides the Alert context to all of
the components that it uses.