Rsdev 684 limit search input to a single molecule (#373)
## Description ##
Chemistry search currently only works with a single molecule as input,
although there was nothing stopping users adding multiple elements. This
change adds some helper text at the top of the search dialog, and also
validates that there's only 1 chemical element on the canvas when the
user submits the search.

Other minor changes to get this to work:
- Moved the code to reset the canvas `window.ketcher.setMolecule("")`,
which was handled in the child Ketcher component and tied to clicking
the submit button, to each of the parents, as in the case of search we
don't always want to clear the canvas, as the search input might not be
valid. This way the user has a chance to fix the issue without having to
recreate their search input
- Wrapped `KetcherTinyMCE` in a `<ThemeProvider>` as there was an error
around themes after adding the `ValidatingSubmitButton` to the child
component

There's also an auto-formatting change in some unrelated java code.

After these changes, this is what the user sees when they try to search
with more than 1 element:

<img width="1461" alt="Screenshot 2025-06-17 at 11 32 30"
src="https://github.com/user-attachments/assets/0e7efcaf-72c2-48e1-b73a-6ed57cfd8f91"
/>

---------

Co-authored-by: rs-fraser <fraser.macinnes>
Co-authored-by: Robert Lamacraft <robert.lamacraft@researchspace.com>