Rsdev 682 websocket backend (#357)
## Description ##
This change adds a simple message broker where a client can subscribe to
a topic at `/topic/notifications/{userId}` and receive an updated
notification count when the user has a new message.

The format of the message is the same as the existing polled message:
```
{
 "notificationCount": 15,
 "messageCount": 42,
 "specialMessageCount": 7
}
```

There's example code for how to connect in this poc branch:
https://github.com/ResearchSpace-ELN/rspace-web/tree/rsdev-679-backend-websockets-poc

A message is triggered each time a message or notification count changes
inside `MessageAndNotificationTracker`.

## Testing notes
This change can be seen in action in the POC branch, which has these
changes plus the front-end modified to use websockets to update
notificaitions.

---------

Co-authored-by: rs-fraser <fraser.macinnes>
Co-authored-by: Robert Lamacraft <robert.lamacraft@researchspace.com>