RSDEV-700 Rendering error analytics event (#388)
## Description ##
This change adds an analytics event that gets logged when react fails to
render a page. This gives us some insight into UI instability that
should allow to us to detect and resolve bugs proactively rather than
being reliant to reports from users.

This change also involves a little bit of refactoring of the analytics
context code

## Testing notes
Testing is difficult because we don't have any code that intentionally
crashes the rendering logic


This change depends on the final piece of TypeScript migration, so as to
not create conflict, and will be ready for release once that change has
been merged.