Revert Analytics refactor (#400)
This change reverts a recent refactoring of the Analytics component
because it broke some of the workspace's create menu items. Ultimately I
think this is because for some very strange reason we have jQuery
messing with DOM nodes created by react, but now is not the time to
re-write all of that.