PRT-955 Downgrade Ketcher to avoid error on opening. (#408)
## Description ##
There was an error when trying to open Ketcher for a 2nd time after
having previously opened it.

Seems to be caused by the upgrade to version 3.4 as part of periodic
front-end dependency upgrades.

I've downgraded Ketcher to the version we previously used and we can
handle the upgrade as a separate task.

---------

Co-authored-by: rs-fraser <fraser.macinnes>