Remove hidden heading from apps page app bar (#412)
## Description ##
The apps page doesn't need a second, hidden, h1 as it already has one
and it would be an accessibility violation to have two. The principle in
theory applies to all other pages that do not appear as tabs in the app
bar as they too should have titles for sighted users as there is no
indication in the app bar what page the user is currently on.