Adjust Inventory page title (#413)
## Description ##
The page title should update as the user navigates around a SPA so that
the user can use standard browser features, like tab names, bookmarking
defaults, and the browser's history just as if they were using a
traditional multi-page site.

This is to ensure compliance with [WCAG 2.2 Criteria 2.4.2
(A)](https://www.w3.org/TR/WCAG22/#page-titled).