RSDEV-738 Adjust the Call-to-action colour to ensure there is sufficient contrast (#417)
## Description ##
We use a blue colour across the product to make certain UI elements
stand out, to call attention to the fact that there are the actions that
the user probably wants to take: the submission of forms, actioning a
selection of items, creating new items, etc.

This background blue colour, coupled with the white text, did not have
sufficient contrast to meet the WCAG 2.2 standard to the required AA
level so this change adjusts the colours, opting for a more saturated
background blue with darker text.