RSDEV-757 Unit selector should always come after the numerical value (#426)
## Description ##
This change just fixes a small inconsistency between the quantity fields
in the subsample editor form and the new sample form, where the quantity
selector came before the numerical field in the former and after in the
latter. This change moves the subsample form's unit selector to after
the numerical field to make these fields consistent.