RSDEV-520: Extend Fieldmark import with IGSN ID (#418)
This commit PR extends the functionality "Import from Fieldmark"

From now when importing from fieldmark the system will analyse all the rows and returns a list of candidate fields that can be used as DOI identifiers. Finally the user will be able select one of the candidate fields to be treated as IGSN DOI Identifier. Finally, when importing the fieldmark notebook into the inventory, the Samples will already have the DOI linked as specified on the Fieldmark UI.