Rsdev 575 add ascenscia (#376)
## Description ##
This PR adds an Ascenscia app controller, with a `/connect` endpoint.

The user enters their Ascenscia username, password and organization on
the apps page entry, and via the back-end an Ascenscia api token is
generated and stored in the db.

This will allow the Ascenscia integration to get the api key by calling
`/integration/integrationInfo?name=ASCENSCIA`.

Where there's an error at the Ascenscia end, we try to relay the error
as when the password or org is incorrect, Ascenscia includes that
information. In other cases, including when the username is wrong, we
return a general 500 error as there's no error message to relay from
Ascenscia.

There's also validation on the back-end to ensure each field is not
blank with appropriate error messages.

In the previous commit
[commit](https://github.com/rspace-os/rspace-web/pull/376/commits/2ae626ebc6133ae2a9198f5d9fcd73def7dcd051)
there's a button on the apps page UI I used during testing to show the
token, which might be useful if testing locally.

I've removed the refresh token code for now as I realised earlier we
need to store the Ascenscia username and org to be able to use the
refresh endpoint, which I'd overlooked as I was using hardcoded values,
so I'll pick that up via a new ticket. If when using the Ascenscia
integration the token stops working, the user should be directed back to
the apps page to generate a new one.

## Testing
Flow:
- sysadmin enables deployment property `ascenscia.enabled`
- user enables Ascnescia app on apps page
- user enters Ascenscia username, password and organisation on the apps
page
    - these credentials aren't stored by rspace
- `rspace-web` backend calls Ascenscia api and swaps credentials for JWT
auth and refresh tokens and stores both in the db
- user is notified auth token was successful, or relays error from
Ascenscia api if there's a problem
- the Ascenscia integration in the editor gets the users Ascenscia auth
token via `/integration/integrationInfo?name=ASCENSCIA`
- ~~if there's an auth error when using the token, Ascenscia integration
calls `rspace-web` `/refresh` endpoint to refresh the token~~ (refresh
not implemented yet)
- ~~if the refresh is unsuccessful,~~ Ascenscia integration should tell
the user to enter their credentials again and generate a new token via
the rspace-web apps page

---------

Co-authored-by: rs-fraser <fraser.macinnes>
Co-authored-by: Robert Lamacraft <robert.lamacraft@researchspace.com>